backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.12.0
numpy>=1.23.0
Pillow>=9.5.0